+++ app/abilities/dsj/event_ability.rb
@@ -1,0 +1,22 @@
+# frozen_string_literal: true
+
+#  Copyright (c) 2012-2025, Dachverband Schweizer Jugendparlamente. This file is part of
+#  hitobito_dsj and licensed under the Affero General Public License version 3
+#  or later. See the COPYING file at the top-level directory or at
+#  https://github.com/hitobito/hitobito_dsj.
+
+module Dsj::EventAbility
+  extend ActiveSupport::Concern
+
+  included do
+    on(Event::Fundraising) do
+      permission(:group_full).may(:list_fundraisings).in_same_group
+
+      permission(:group_and_below_full).may(:list_fundraisings).in_same_group_or_below
+
+      permission(:layer_full).may(:list_fundraisings).in_same_layer
+
+      permission(:layer_and_below_full).may(:list_fundraisings).in_same_layer_or_below
+    end
+  end
+end