+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Hitobito Changelog
 
+## unreleased
+
+*   Neuer Event Typ Fundraising (hitobito/hitobito_dsj#34)
+
 ## Version 1.13
 
 *   Zusätzliche Felder auf Gruppe