+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Hitobito Changelog
 
+## unreleased
+
+*   Neu können Fundraisings mit Förder:innen in hitobito erfasst und verwaltet werden (hitobito/hitobito_dsj#34)
+
 ## Version 1.13
 
 *   Zusätzliche Felder auf Gruppe